commitAuth1

+++ src/Untappd.Net/Responses/BeerInfo.cs
@@ -159,7 +159,7 @@ namespace Untappd.Net.Responses.BeerInfo
         public string BeerStyle { get; set; }
 
         [JsonProperty("auth_rating")]
-        public int AuthRating { get; set; }
+        public double AuthRating { get; set; }
 
         [JsonProperty("wish_list")]
         public bool WishList { get; set; }
@@ -362,7 +362,7 @@ namespace Untappd.Net.Responses.BeerInfo
         public string BeerStyle { get; set; }
 
         [JsonProperty("auth_rating")]
-        public int AuthRating { get; set; }
+        public double AuthRating { get; set; }
 
         [JsonProperty("wish_list")]
         public bool WishList { get; set; }
@@ -623,7 +623,7 @@ namespace Untappd.Net.Responses.BeerInfo
         public string BeerLabel { get; set; }
 
         [JsonProperty("auth_rating")]
-        public int AuthRating { get; set; }
+        public double AuthRating { get; set; }
 
         [JsonProperty("wish_list")]
         public bool WishList { get; set; }
@@ -823,7 +823,7 @@ namespace Untappd.Net.Responses.BeerInfo
         public Brewery Brewery { get; set; }
 
         [JsonProperty("auth_rating")]
-        public int AuthRating { get; set; }
+        public double AuthRating { get; set; }
 
         [JsonProperty("wish_list")]
         public bool WishList { get; set; }
