+++ src/Untappd.Net/Responses/Feeds/UserActivityFeed.cs
@@ -123,7 +123,7 @@ namespace Untappd.Net.Responses.Feeds
         public double BeerAbv { get; set; }
 
         [JsonProperty("auth_rating")]
-        public int AuthRating { get; set; }
+        public double AuthRating { get; set; }
 
         [JsonProperty("wish_list")]
         public bool WishList { get; set; }
